Update reference files at summit.
The translator test reference files have been regenerated
at the summit.  The translator output is different
at the summit because the configuration under
/jac_sw/itsroot must be different between the summit
and Hilo.  Using the summit version means that the
translator test uses the real configuration,
and that the reference files are also suitable for
testing the live version of the translator.